applyTripReplacementPeriods: handle start_date in a timezone-independent way 🐛
Since the `date` column created by gtfs-via-postgres is a `timestamp without time zone`, and since the `pg` client converts it to a ES `Date` by default, we cannot use `Date`'s timezone-naive getters.

Because we want the unmodified `date` value – it is used in `start_date` to identify the respective train/"run" of a trip, after all –, we convert it to a ISO 8601 *string* in SQL already.